"Finished" the vm, script loader, and script initialization. Try out src/main.c to run Game::play(), which is used to start up SCI games (AFAIK). Everything is implemented (but not thoroughly tested) except for the kernel functions. It's somewhat of a mess, but it does things quite similar to the way they're supposed to be done, so it'll be a starting point to get the thing moving.
Try src/main.c (compiled into an executable called 'sciv' for no particular
reason) to see it move (uses the default command console functionality with
readline() as the input function).

Available debugger functionality:
 - registers : dumps all current register values
 - s : execute an arbitrary number of operations (defaulting to 1)
 - heapdump : dump heap data
 - disasm : Disassemble arbitrary parts of the heap
 - scriptinfo : Show the table of loaded scripts
 - heapobj : Determine whether an object or class is at the specified heap
   address, and, if it is, dump all selector names and values for it.
 - classtable : Dump the table of all classes
 - stack : Dump stack elements
 - bt : Show execution stack affected by (send, self, super, call, calle,
   callb, ret)

More are to come.

Expect 0.2.3 to be released early next week.

 - Christoph

+++ src/core/heap.c
@@ -32,16 +32,16 @@ static void set_next(heap_t* h, int block_pos, int next)
 	return src[0]|((src[1])<<8);
 	}*/
 
-static int get_size(heap_t* h, int block_pos)
+static unsigned int get_size(heap_t* h, int block_pos)
 {
 	assert_in_range(block_pos);
-	return getInt16(h->start+block_pos);
+	return (guint16)getInt16(h->start+block_pos);
 }
 
-static int get_next(heap_t* h, int block_pos)
+static unsigned int get_next(heap_t* h, int block_pos)
 {
 	assert_in_range(block_pos);
-	return getInt16(h->start+block_pos+2);
+	return (guint16)getInt16(h->start+block_pos+2);
 }
 
 /*Allocates a new heap*/
@@ -228,6 +228,8 @@ void heap_dump_free(heap_t *h)
 }
 
 
+/*
+
 int main(int argc, char **argv) {
   heap_t *h = heap_new();
   int a,b,c,d,e;
@@ -261,3 +263,4 @@ int main(int argc, char **argv) {
   return 0;
 }
 
+*/
